app/requirements.txt
Flask==2.3.3
openai==0.28.1
pypdf==4.0.0
python-dotenv==1.0.0
reportlab==4.0.7
Werkzeug==2.3.7
